benchmarks/pyproject.toml
[project]
name = "benchmarks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
